src/requirements.txt
matplotlib==3.8.0
nltk==3.8.1
numpy==1.26.4
pandas==2.0.0
Pillow==10.2.0
scikit_learn==1.3.2
streamlit==1.32.2
streamlit_option_menu==0.3.12
torch==2.2.1
transformers==4.35.2
